src/requirements.txt
torch
torchtext
numpy
typing-extensions
sacrebleu
transformers
tensorflow-datasets
tensorflow-text
tensorflow-metal